Python 3 aula10 condições parte 1 exercícios

+++ exercicios/ex035.py
@@ -1,6 +1,6 @@
-a = int(input('Digite a reta a: '))
-b = int(input('Digite a reta b: '))
-c = int(input('Digite a reta c: '))
+a = float(input('Digite a reta a: '))
+b = float(input('Digite a reta b: '))
+c = float(input('Digite a reta c: '))
 if a < b + c and b < a + c and c < a + b:
     print(f'As retas a = {a} e b = {b} e c = {c} formam um triângulo.')
 else:
